+++ models/BaseMongoEntityModel.php
@@ -33,13 +33,14 @@ abstract class BaseMongoEntityModel extends \yii\mongodb\ActiveRecord
         if ($this->skipInit) {
             return;
         }
+        $this->idAttribute = '_id';
         $this->initEntityEvents();
         parent::init();
     }
 
     /**
      * @inheritdoc
-     * @return \vistart\Models\models\BaseEntityQuery the newly created [[BaseEntityQuery]] or its sub-class instance.
+     * @return \vistart\Models\queries\BaseMongoEntityQuery the newly created [[BaseEntityQuery]] or its sub-class instance.
      */
     public static function find()
     {
@@ -49,5 +50,10 @@ abstract class BaseMongoEntityModel extends \yii\mongodb\ActiveRecord
         }
         $queryClass = $self->queryClass;
         return new $queryClass(get_called_class(), ['noInitModel' => $self]);
+    }
+
+    public function attributes()
+    {
+        return $this->enabledFields();
     }
 }