+++ tests/TestCase.php
@@ -104,11 +104,12 @@ abstract class TestCase extends \PHPUnit_Framework_TestCase {
             $this->markTestSkipped('No mysql server connection configured.');
         }
         $connection = new Connection($params);
+        $redis = self::getParam('redis');
         $cacheParams = self::getParam('cache');/*
         if ($cacheParams === null) {
             $this->markTestSkipped('No cache component configured.');;
         }*/
-        $this->mockWebApplication(['components' => ['db' => $connection, 'cache' => $cacheParams]]);
+        $this->mockWebApplication(['components' => ['redis' => $redis, 'db' => $connection, 'cache' => $cacheParams]]);
 
         parent::setUp();
     }